requirements.txt
# Core dependencies
fastapi>=0.104.1,<0.112.0
uvicorn[standard]>=0.24.0,<0.31.0
openai>=1.13.0,<2.0.0
langchain>=0.1.0,<0.3.0
langchain-openai>=0.0.8,<0.2.0
chromadb>=0.4.18,<0.6.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
sqlalchemy>=2.0.23,<3.0.0
alembic>=1.13.0,<2.0.0
python-multipart>=0.0.6,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=1.0.0,<2.0.0
loguru>=0.7.2,<1.0.0
httpx>=0.25.0,<1.0.0
aiofiles>=23.2.1,<24.0.0
jinja2>=3.1.2,<4.0.0

# Development dependencies
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.21.1,<1.0.0
pytest-cov>=4.1.0,<5.0.0
black>=23.11.0,<25.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<8.0.0
mypy>=1.7.1,<2.0.0
pre-commit>=3.6.0,<4.0.0

# Optional dependencies for enhanced features
sentence-transformers>=2.2.2,<3.0.0
numpy>=1.24.4,<2.0.0
pandas>=2.1.4,<3.0.0
scikit-learn>=1.3.2,<2.0.0
